requirements.txt
Flask
requests
python-dotenv
pymongo[srv]
